Move `GraphManager` to `graphs` module
