--- docs/api/ford.graphmanager.rst
@@ -1,7 +1,0 @@
-ford.graphmanager module
-========================
-
-.. automodule:: ford.graphmanager
-   :members:
-   :undoc-members:
-   :show-inheritance:
